+++ packages/expo-image-manipulator/src/ImageManipulator.types.ts
@@ -1,4 +1,5 @@
 import type { NativeModule } from 'expo';
+import { SharedRef } from 'expo-modules-core/types';
 
 import type { ImageManipulatorContext } from './ImageManipulatorContext';
 import ImageRef from './ImageRef';
@@ -138,5 +139,5 @@ export declare class ImageManipulator extends NativeModule {
   /**
    * Loads an image from the given URI and creates a new image manipulation context.
    */
-  manipulate(uri: string): ImageManipulatorContext;
+  manipulate(source: string | SharedRef<'image'>): ImageManipulatorContext;
 }